Close Formula Keyboard definition invariants
- enforce definition validation across decoding paths
- finalize identifier validation semantics
- normalize integer layout weight invariants
- add invalid decoding coverage

+++ EMathicaFormulaKeyboardKit/Sources/EMathicaFormulaKeyboardCore/Validation/FormulaKeyboardDefinitionError.swift
@@ -10,6 +10,6 @@ public enum FormulaKeyboardDefinitionError: Error, Hashable, Equatable, Codable,
     case duplicateRowIdentifier(FormulaKeyboardRowIdentifier)
     case duplicateKeyIdentifier(FormulaKeyIdentifier)
     case invalidIdentifier(String)
-    case invalidWidthWeight(Double)
+    case invalidWidthWeight(Int)
     case missingDefaultPage(FormulaKeyboardPageIdentifier)
 }
